dont override existing config file

+++ INSTALL.sh
@@ -9,8 +9,13 @@ if [ ! -d $config_dir ]; then
     mkdir $config_dir
 fi
 
-cp config $config_dir/
-cp img/coffee.png $config_dir/
+if [ ! -f "$config_dir/config" ]; then
+    cp config $config_dir/
+fi
+
+if [ ! -f "$config_dir/coffee.png" ]; then
+    cp img/coffee.png $config_dir/
+fi
 
 echo "start deamon: eyeronic --config=~/.config/eyeronic/config --start"
